requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.15.1
idna==3.7
oauthlib==3.2.2
pillow==10.3.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
urllib3==2.2.2
